Remove -I. from linux and posix builds
